jarvan/feature-environment:更新数据库；修改实体类；测试插入部分数据

+++ leetcode/src/main/java/com/leetcode/web/entity/AnswersTags.java
@@ -30,7 +30,7 @@ public class AnswersTags implements Serializable {
      * 题解的id
      */
     @TableId
-    private Integer answerId;
+    private Long answerId;
 
     /**
      * 标签的id
